refactor: review request handler, agent and agent base
