+++ contract-agent/src/tests/mocks/contract-agent.config.json
@@ -3,12 +3,15 @@
     {
       "source": "contracts",
       "url": "mongodb://localhost:27017",
-      "dbName": "contract_consent_agent_db_test"
+      "dbName": "contract_consent_agent_db_test",
+      "watchChanges": true
     },
     {
       "source": "profiles",
       "url": "mongodb://localhost:27017",
-      "dbName": "contract_consent_agent_db_test"
+      "dbName": "contract_consent_agent_db_test",
+      "watchChanges": false,
+      "hostsProfiles": true
     }
   ]
 }
